Many files: [[Header/lint cleanup]]
* ntpdate/ntptimeset.c:
* ntpdate/ntpdate.h:
* ntpd/refclock_oncore.c (oncore_msg_En):
* ntpd/ntp_util.c (stats_config):
* ntpd/ntp_request.c:
* ntpd/ntp_intres.c (findhostaddr):
* ntpd/ntp_config.c (getconfig):
* libntp/systime.c (adj_systime):
* libntp/lib_strbuf.c:
* libntp/authparity.c:
* libntp/audio.c:  Header/lint cleanup
  From/reported by: Simon Burge <simonb@netbsd.org>

+++ libntp/systime.c
@@ -159,8 +159,8 @@ adj_systime(
 	/* casey - we need a posix type thang here */
 	if (adjtime(&adjtv, &oadjtv) < 0)
 	{
-		msyslog(LOG_ERR, "Can't adjust time (%d sec, %d usec): %m",
-			adjtv.tv_sec, adjtv.tv_usec);
+		msyslog(LOG_ERR, "Can't adjust time (%ld sec, %ld usec): %m",
+			(long)adjtv.tv_sec, (long)adjtv.tv_usec);
 		return 0;
 	} 
 	else {
