+++ ntpd/refclock_oncore.c
@@ -1587,7 +1587,8 @@ oncore_msg_En(
 
 		if (debug > 2)
 			printf("ONCORE: serial/j (%lu, %d) %ld.%09ld\n",
-			    pps_i.assert_sequence, j, tsp->tv_sec, tsp->tv_nsec);
+			       (long)pps_i.assert_sequence, j,
+			       (long)tsp->tv_sec, (long)tsp->tv_nsec);
 
 		if (pps_i.assert_sequence == j) {
 			printf("ONCORE: oncore_msg_En, error serial pps\n");
